fix(pi): Address PR comments - evidence a build cannot honour is refused
The completion record and a registration's matcher went into the lock
without moving its version, so an older build would have loaded the file,
ignored both, and dropped them on its next write — and then read a
finished move as unfinished and taken back what the person had since put
under the reserved name. The version moves to 5, which is what makes an
older build refuse the file instead. A refusal is visible and
recoverable; a silent drop destroys their bytes.

And the record was being written onto entries that did not exist yet: a
hook nothing declares any more has no entry until the sweep carries its
old one forward, so a pass that finished such a hook's move earned the
fact and then threw it away. The move hands back what it finished and one
writer stamps it after the sweep, where every entry that will exist does.
